Update visualize.Rmd
changed "visualize your data ggplot2" to "visualize your data with ggplot2".

also made a changed "If you run this code and get the error there is no package called ‘ggplot2’" to "If you run this code and get an error there is no package called ‘ggplot2’ on your machine". I'm not sure if "machine" is the correct word to use.