Update intro.Rmd
Typo. 

I'm not if I once seen a reference to Jenny's reprex example here but I was expecting to see a link to her github page that shows how to do the reprex.

Under the Acknowledgements section, were you thank everyone who contributed via GitHub, you seem to be pulling the username and full name (Jennifer (Jenny) Bryan, @jennybc) for the contributors.