Fix a few minor typos
Examples include saying we'll learn about filter() in the next chapter, but it's not until chapter 5.  And, also, it says we'll learn about how to use multiple geoms in the same plot in the next section, but then not far below, in the same section, it shows how to use multiple geoms.