Changes to logicals.qmd (#1234)
* Update EDA.qmd

Using the "size" argument prints the warning: "#> Warning: Using `size` aesthetic for lines was deprecated in ggplot2 3.4.0.
#> ℹ Please use `linewidth` instead."

* Update communication.qmd

Line 844: added *to* to the sentence "It is generally more useful if you want to *expand* the limits..."

Line 1101: changed the legend. Position argument to "top" to align with the statement on line 1053.

* Update logicals.qmd

because "x" does not have any value > 3, both evaluations above are not TRUE. Either the second expression use "x > 2" or "x >= 3".