Improve description of readr heuristics on data-import.qmd
I read "1,000^2" and it wasn't obvious that the "^2" was a citation, and instead I thought that it was an exponent. It was confusing, and it was only when I hovered over the 2 that I realized what was happening. To solve this, I simply moved the "^2" to the next word, so that the "2" is commenting on the broader idea of "1,000 rows".